user-service/pyproject.toml
[tool.poetry]
name = "user-service"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
sqlmodel = "^0.0.22"
python-jose = "^3.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
uvicorn = "^0.32.0"
psycopg = "^3.2.3"
psycopg-binary = "^3.2.3"
python-multipart = "^0.0.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"